Remove numpy linear system.
