+++ src/energnn/problem/example/linear_system.py
@@ -7,11 +7,12 @@
 import copy
 from copy import deepcopy
 
+import jax.numpy as jnp
 import numpy as np
 from omegaconf import DictConfig
 
-from energnn.graph import Graph, GraphShape, GraphStructure, HyperEdgeSet, HyperEdgeSetStructure, collate_graphs
-from energnn.graph.jax import JaxGraph
+from energnn.graph import GraphStructure, HyperEdgeSetStructure
+from energnn.graph.jax import JaxGraph, JaxGraphShape, JaxHyperEdgeSet, collate_graphs_jax
 from ..batch import ProblemBatch
 from ..loader import ProblemLoader
 from ..problem import Problem
@@ -30,16 +31,14 @@ LINEAR_SYSTEM_DECISION_STRUCTURE = GraphStructure(
 class LinearSystemProblemBatch(ProblemBatch):
     __test__ = False
 
-    def __init__(self, *, context: Graph, oracle: Graph):
+    def __init__(self, *, context: JaxGraph, oracle: JaxGraph):
         self.context = context
         self.oracle = oracle
-        self.jax_context = JaxGraph.from_numpy_graph(context)
-        self.jax_oracle = JaxGraph.from_numpy_graph(oracle)
 
         zero_decision = copy.deepcopy(oracle)
+        # Vérifier opération
         zero_decision.feature_flat_array = 0.0 * zero_decision.feature_flat_array
         self.zero_decision = zero_decision
-        self.jax_zero_decision = JaxGraph.from_numpy_graph(zero_decision)
 
     @property
     def decision_structure(self) -> GraphStructure:
@@ -51,32 +50,34 @@ class LinearSystemProblemBatch(ProblemBatch):
 
     def get_context(self, get_info: bool = False, step: int | None = None) -> tuple[JaxGraph, dict]:
         """Returns the context :class:`Graph` :math:`x`."""
-        return deepcopy(self.jax_context), {}
+        return deepcopy(self.context), {}
 
     def get_oracle(self, get_info: bool = False) -> tuple[JaxGraph, dict]:
         r"""Returns the ground truth :class:`Graph` :math:`y^{\star}(x)`."""
-        return deepcopy(self.jax_oracle), {}
+        return deepcopy(self.oracle), {}
 
     def get_zero_decision(self, get_info: bool = False) -> tuple[JaxGraph, dict]:
         """Returns a decision filled with zeros."""
-        return deepcopy(self.jax_zero_decision), {}
+        return deepcopy(self.zero_decision), {}
 
     def get_gradient(
         self, decision: JaxGraph, cfg: DictConfig | None = None, get_info: bool = False, step: int | None = None
-    ) -> tuple[Graph, dict]:
+    ) -> tuple[JaxGraph, dict]:
         r"""Returns the gradient :class:`Graph` :math:`\nabla_y f(y;x) = y - y^{\star}(x)`."""
-        gradient = decision.to_numpy_graph()
+        # gradient = decision.to_numpy_graph()
+        gradient = deepcopy(decision)
         gradient.feature_flat_array = gradient.feature_flat_array - self.oracle.feature_flat_array
-        jax_gradient = JaxGraph.from_numpy_graph(gradient)
-        return jax_gradient, {}
+        # jax_gradient = JaxGraph.from_numpy_graph(gradient)
+        return gradient, {}
 
     def get_score(
         self, decision: JaxGraph, cfg: DictConfig | None = None, get_info: bool = False, step: int | None = None
     ) -> tuple[list[float], dict]:
         """Returns the mean-squared error of the decision :class:`Graph` with regard to the oracle :class:`Graph`."""
-        gradient = decision.to_numpy_graph()
+        # gradient = decision.to_numpy_graph()
+        gradient = deepcopy(decision)
         gradient.feature_flat_array = gradient.feature_flat_array - self.oracle.feature_flat_array
-        objective = np.nanmean(np.square(gradient.feature_flat_array), axis=1)
+        objective = jnp.nanmean(jnp.square(gradient.feature_flat_array), axis=1)
         return objective.tolist(), {}
 
     def save(self, *, path: str) -> None:
@@ -86,16 +87,13 @@ class LinearSystemProblemBatch(ProblemBatch):
 class LinearSystemProblem(Problem):
     __test__ = False
 
-    def __init__(self, *, context: Graph, oracle: Graph):
+    def __init__(self, *, context: JaxGraph, oracle: JaxGraph):
         self.context = context
         self.oracle = oracle
-        self.jax_context = JaxGraph.from_numpy_graph(context)
-        self.jax_oracle = JaxGraph.from_numpy_graph(oracle)
 
         zero_decision = copy.deepcopy(oracle)
         zero_decision.feature_flat_array = 0.0 * zero_decision.feature_flat_array
         self.zero_decision = zero_decision
-        self.jax_zero_decision = JaxGraph.from_numpy_graph(zero_decision)
 
     @property
     def decision_structure(self) -> GraphStructure:
@@ -107,32 +105,34 @@ class LinearSystemProblem(Problem):
 
     def get_context(self, get_info: bool = False, step: int | None = None) -> tuple[JaxGraph, dict]:
         """Returns the context :class:`Graph` :math:`x`."""
-        return deepcopy(self.jax_context), {}
+        return deepcopy(self.context), {}
 
     def get_oracle(self, get_info: bool = False) -> tuple[JaxGraph, dict]:
         r"""Returns the ground truth :class:`Graph` :math:`y^{\star}(x)`."""
-        return deepcopy(self.jax_oracle), {}
+        return deepcopy(self.oracle), {}
 
     def get_zero_decision(self, get_info: bool = False) -> tuple[JaxGraph, dict]:
         """Returns a decision filled with zeros."""
-        return deepcopy(self.jax_zero_decision), {}
+        return deepcopy(self.zero_decision), {}
 
     def get_gradient(
         self, decision: JaxGraph, cfg: DictConfig | None = None, get_info: bool = False, step: int | None = None
     ) -> tuple[JaxGraph, dict]:
         r"""Returns the gradient :class:`Graph` :math:`\nabla_y f(y;x) = y - y^{\star}(x)`."""
-        gradient = decision.to_numpy_graph()
+        # gradient = decision.to_numpy_graph()
+        gradient = deepcopy(decision)
         gradient.feature_flat_array = gradient.feature_flat_array - self.oracle.feature_flat_array
-        jax_gradient = JaxGraph.from_numpy_graph(gradient)
-        return jax_gradient, {}
+        # jax_gradient = JaxGraph.from_numpy_graph(gradient)
+        return gradient, {}
 
     def get_score(
         self, decision: JaxGraph, cfg: DictConfig | None = None, get_info: bool = False, step: int | None = None
     ) -> tuple[float, dict]:
         """Returns the mean-squared error of the decision :class:`Graph` with regard to the oracle :class:`Graph`."""
-        gradient = decision.to_numpy_graph()
+        # gradient = decision.to_numpy_graph()
+        gradient = deepcopy(decision)
         gradient.feature_flat_array = gradient.feature_flat_array - self.oracle.feature_flat_array
-        objective = np.nanmean(np.square(gradient.feature_flat_array))
+        objective = jnp.nanmean(jnp.square(gradient.feature_flat_array))
         return float(objective), {}
 
     def save(self, *, path: str) -> None:
@@ -166,15 +166,15 @@ class LinearSystemProblemGenerator:
         A, b, x = _generate_sparse_linear_system(n, m)
 
         # Context
-        arrow_edge = HyperEdgeSet.from_dict(
+        arrow_edge = JaxHyperEdgeSet.from_dict(
             port_dict={"from": np.nonzero(A)[0], "to": np.nonzero(A)[1]}, feature_dict={"value": A[np.nonzero(A)]}
         )
-        source_edge = HyperEdgeSet.from_dict(port_dict={"id": np.arange(n)}, feature_dict={"value": b})
-        context = Graph.from_dict(hyper_edge_set_dict={"arrow": arrow_edge, "source": source_edge}, n_addresses=n)
+        source_edge = JaxHyperEdgeSet.from_dict(port_dict={"id": np.arange(n)}, feature_dict={"value": b})
+        context = JaxGraph.from_dict(hyper_edge_set_dict={"arrow": arrow_edge, "source": source_edge}, n_addresses=n)
 
         # Oracle
-        source_edge = HyperEdgeSet.from_dict(port_dict=None, feature_dict={"value": x})
-        oracle = Graph.from_dict(hyper_edge_set_dict={"source": source_edge}, n_addresses=n)
+        source_edge = JaxHyperEdgeSet.from_dict(port_dict=None, feature_dict={"value": x})
+        oracle = JaxGraph.from_dict(hyper_edge_set_dict={"source": source_edge}, n_addresses=n)
 
         return LinearSystemProblem(context=context, oracle=oracle)
 
@@ -189,15 +189,16 @@ class LinearSystemProblemGenerator:
             context_list.append(context)
             oracle_list.append(oracle)
 
-        max_context_shape = GraphShape(
-            hyper_edge_sets={"arrow": np.array(self.n_max**2), "source": np.array(self.n_max)}, addresses=np.array(self.n_max)
+        max_context_shape = JaxGraphShape(
+            hyper_edge_sets={"arrow": jnp.array(self.n_max**2), "source": jnp.array(self.n_max)},
+            addresses=jnp.array(self.n_max),
         )
-        max_oracle_shape = GraphShape(hyper_edge_sets={"source": np.array(self.n_max)}, addresses=np.array(self.n_max))
+        max_oracle_shape = JaxGraphShape(hyper_edge_sets={"source": jnp.array(self.n_max)}, addresses=jnp.array(self.n_max))
 
         [context.pad(target_shape=max_context_shape) for context in context_list]
         [oracle.pad(target_shape=max_oracle_shape) for oracle in oracle_list]
-        context_batch = collate_graphs(context_list)
-        oracle_batch = collate_graphs(oracle_list)
+        context_batch = collate_graphs_jax(context_list)
+        oracle_batch = collate_graphs_jax(oracle_list)
 
         return LinearSystemProblemBatch(context=context_batch, oracle=oracle_batch)
 
@@ -248,31 +249,3 @@ class LinearSystemProblemLoader(ProblemLoader):
 
     def __len__(self):
         return max(self.dataset_size // self.batch_size, 1)
-
-
-def compare_single_graphs(a: JaxGraph, b: JaxGraph, rtol=1e-5, atol=1e-6):
-    """
-    Compare two single (non-batched) JaxGraph objects component-wise.
-    """
-    assert set(a.hyper_edge_sets.keys()) == set(
-        b.hyper_edge_sets.keys()
-    ), f"Edge keys differ: {set(a.hyper_edge_sets.keys())} vs {set(b.hyper_edge_sets.keys())}"
-    for k in a.hyper_edge_sets:
-        ae = a.hyper_edge_sets[k]
-        be = b.hyper_edge_sets[k]
-        # feature arrays
-        if (ae.feature_array is None) != (be.feature_array is None):
-            raise AssertionError(f"Feature presence mismatch for edge {k}")
-        if ae.feature_array is not None:
-            np.testing.assert_allclose(np.array(ae.feature_array), np.array(be.feature_array), rtol=rtol, atol=atol)
-        # address_dict keys
-        a_keys = set(ae.port_dict.keys()) if ae.port_dict is not None else set()
-        b_keys = set(be.port_dict.keys()) if be.port_dict is not None else set()
-        assert a_keys == b_keys
-        for ak in a_keys:
-            np.testing.assert_allclose(np.array(ae.port_dict[ak]), np.array(be.port_dict[ak]), rtol=rtol, atol=atol)
-        # non_fictitious mask
-        if ae.non_fictitious is None or be.non_fictitious is None:
-            assert ae.non_fictitious is be.non_fictitious
-        else:
-            np.testing.assert_allclose(np.array(ae.non_fictitious), np.array(be.non_fictitious), rtol=rtol, atol=atol)
